Fix for the naming of the Java file used to store a code generated quote

+++ core/codegen/src/main/java/org/overture/codegen/utils/GeneralUtils.java
@@ -138,13 +138,13 @@ public class GeneralUtils
 		return files;
 	}
 
-	public static void deleteFolderContents(File folder)
+	public static void deleteFolderContents(File folder, boolean removeFolders)
 	{
-		deleteFolderContents(folder, new ArrayList<String>());
+		deleteFolderContents(folder, new ArrayList<String>(), removeFolders);
 	}
 
 	public static void deleteFolderContents(File folder,
-			List<String> folderNamesToAvoid)
+			List<String> folderNamesToAvoid, boolean removeFolders)
 	{
 		if (folder == null)
 		{
@@ -164,9 +164,14 @@ public class GeneralUtils
 			{
 				if (!folderNamesToAvoid.contains(f.getName()))
 				{
-					deleteFolderContents(f, folderNamesToAvoid);
+					deleteFolderContents(f, folderNamesToAvoid, removeFolders);
 				}
 			} else
+			{
+				f.delete();
+			}
+			
+			if(removeFolders)
 			{
 				f.delete();
 			}
